ai-service/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
numpy==1.26.4
pydantic==2.6.1
python-dotenv==1.0.1
pytest==8.0.0
hypothesis==6.98.3
httpx==0.26.0
